+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Cast Assert Equal
+- Count Assert Equal
 - [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Fixed